[rs] Update wgpu with push constants API

+++ wgpu/src/backend/web.rs
@@ -109,7 +109,7 @@ impl crate::ComputePassInner<Context> for ComputePass {
                 offsets.len() as u32,
             );
     }
-    fn set_push_constants(&mut self, _offset: u32, _data: &[u32]) {
+    fn set_push_constants(&mut self, _offset: u32, _data: &[u8]) {
         panic!("PUSH_CONSTANTS feature must be enabled to call multi_draw_indexed_indirect")
     }
 
@@ -184,7 +184,7 @@ impl crate::RenderInner<Context> for RenderPass {
         self.0
             .set_vertex_buffer_with_f64_and_f64(slot, &buffer.0, offset as f64, mapped_size);
     }
-    fn set_push_constants(&mut self, _stages: wgt::ShaderStage, _offset: u32, _data: &[u32]) {
+    fn set_push_constants(&mut self, _stages: wgt::ShaderStage, _offset: u32, _data: &[u8]) {
         panic!("PUSH_CONSTANTS feature must be enabled to call multi_draw_indexed_indirect")
     }
     fn draw(&mut self, vertices: Range<u32>, instances: Range<u32>) {
